+++ src/main/java/uz/pdp/controller/AuthController.java
@@ -12,8 +12,6 @@ import uz.pdp.enumerators.UserRole;
 import uz.pdp.service.UserService;
 import uz.pdp.service.VerificationService;
 
-import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Controller
 @RequestMapping("/auth")
@@ -46,31 +44,6 @@ public class AuthController {
         return "balance";
     }
 
-    @RequestMapping("/back-balance")
-    public String backBalancePage(Model model, HttpSession session) {
-        User user = (User)session.getAttribute("user");
-        model.addAttribute("user", user);
-        model.addAttribute("balance", user.getBalance());
-        return "patient-page";
-    }
-
-
-    @PostMapping("add-balance")
-    public String addBalance(@RequestParam(name = "amount") double amount, Model model, HttpSession session){
-        User user = (User) session.getAttribute("user");
-        if(amount <= 0){
-            model.addAttribute("message", "Amount must be greater than 0");
-            model.addAttribute("balance", user.getBalance());
-            return "balance";
-        }
-
-        user.setBalance(user.getBalance() + amount);
-        userService.update(user);
-        model.addAttribute("balance", user.getBalance());
-        return "balance";
-    }
-
-
 
 
     @RequestMapping(value = "/login", method = RequestMethod.POST)
@@ -82,7 +55,10 @@ public class AuthController {
         }else if (userEntity.getRole() == UserRole.MAIN_DOCTOR) {
             model.addAttribute("users", userService.getAllDoctors());
             return "admin-page";
-        }else if(userEntity.getRole() != null) {
+        }else if(userEntity.getRole() == UserRole.ADMINISTRATOR){
+            return "admin-page";
+        }
+        else if(userEntity.getRole() != null) {
             return "doctor-page";
         }
         model.addAttribute("error", "Username or password incorrect");
@@ -132,38 +108,6 @@ public class AuthController {
 
     }
 
-
-    @PostMapping("/create-doctors")
-    public String create(@ModelAttribute User userEntity, Model model , HttpSession session) {
-        userEntity.setCreatedAt(LocalDateTime.now());
-        userEntity.setUpdatedAt(LocalDateTime.now());
-        userService.save(userEntity);
-        model.addAttribute("users", userService.getAllDoctors());
-        return "admin-page";
-    }
-
-
-    @GetMapping("create" )
-    public String create(Model model) {
-        model.addAttribute("users", userService.getAllDoctors());
-        return "auth/create";
-    }
-
-    @RequestMapping("/delete-doctor")
-    public String delete(@RequestParam(name = "userId") UUID userId, Model model) {
-        userService.delete(userId);
-        model.addAttribute("users", userService.getAllDoctors());
-        return "admin-page";
-   }
-
-    @RequestMapping(value = "/update-doctor", method = RequestMethod.POST)
-    public String update(@RequestParam(name = "userId") UUID userID , @ModelAttribute User updatedUser, Model model) {
-        updatedUser.setId(userID);
-        userService.update(updatedUser);
-        model.addAttribute("message", "Doctor updated successfully");
-        model.addAttribute("users", userService.getAllDoctors());
-        return "admin-page";
-    }
 
 
 }
