+++ src/main/java/uz/pdp/enumerators/UserRole.java
@@ -5,6 +5,7 @@ public enum UserRole {
     THERAPIST,
     PEDIATRICIAN,
     OTOLARYNGOLOGY,
+    ADMINISTRATOR,
     CARDIOLOGIST,
     RHEUMATOLOGIST,
     ENDOCRINOLOGIST,
