+++ web/WEB-INF/views/patient-page.jsp
@@ -49,7 +49,7 @@
     </button>
 
     <button class="button">
-        <a href="${pageContext.request.contextPath}/auth/balance">Balance</a>
+        <a href="${pageContext.request.contextPath}/admin/balance">Balance</a>
     </button>
 
     <button class="button">
